add message prompting visitors to log in before checking out

+++ app/controllers/invoices_controller.rb
@@ -1,10 +1,14 @@
 class InvoicesController < ApplicationController
   def create
-    customer = Customer.find(current_user.customer_id)
-    invoice = customer.invoices.create
-    invoice.populate(cart.contents)
-    session.delete(:cart)
-    flash[:notice] = "Order Submitted"
+    if current_user
+      customer = Customer.find(current_user.customer_id)
+      invoice = customer.invoices.create
+      invoice.populate(cart.contents)
+      session.delete(:cart)
+      flash[:notice] = "Order Submitted"
+    else
+      flash[:notice] = "Please log in or create a new account"
+    end
     redirect_to root_path
   end
 end
